+++ testlib/src/test/java/com/diffplug/spotless/biome/BiomeStepTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2023-2024 DiffPlug
+ * Copyright 2023-2025 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -53,7 +53,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCjs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
 		}
@@ -64,7 +64,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCts() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.cts", "biome/ts/fileAfter.cts");
 		}
@@ -76,7 +76,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testAutoDetectCssExperimental() {
 			var path = createBiomeConfig("biome/config/css-enabled.json");
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.8.3", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.8.3", downloadDir.toString()).withConfigPath(path).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -87,7 +87,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectCssStable() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.9.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.9.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -98,7 +98,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.js", "biome/js/fileAfter.js");
 		}
@@ -109,7 +109,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJson() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/fileBefore.json", "biome/json/fileAfter.json");
 		}
@@ -120,7 +120,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJsonc() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/jsonc/fileBefore.jsonc", "biome/jsonc/fileAfter.jsonc");
 		}
@@ -131,7 +131,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectJsx() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.jsx", "biome/js/fileAfter.jsx");
 		}
@@ -142,7 +142,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectMjs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.mjs", "biome/js/fileAfter.mjs");
 		}
@@ -153,7 +153,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectMts() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.mts", "biome/ts/fileAfter.mts");
 		}
@@ -164,7 +164,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectTs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.ts", "biome/ts/fileAfter.ts");
 		}
@@ -175,7 +175,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testAutoDetectTsx() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.tsx", "biome/ts/fileAfter.tsx");
 		}
@@ -188,9 +188,35 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void preservesIgnoredFiles() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.5.0", downloadDir.toString()).create();
+			var step = BiomeStep.withExeDownload("1.5.0", downloadDir.toString()).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/package.json", "biome/json/packageAfter.json");
+		}
+	}
+
+	/**
+	 * Tests for the different biome versions.
+	 */
+	@Nested
+	class BiomeVersion {
+		/**
+		 * Basic test that biome 1.x works.
+		 */
+		@Test
+		void test1x() {
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir).create();
+			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
+			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
+		}
+
+		/**
+		 * Basic test that biome 2.x works.
+		 */
+		@Test
+		void test2x() {
+			var step = BiomeStep.withExeDownload("2.0.6", downloadDir).create();
+			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
+			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
 		}
 	}
 
@@ -202,7 +228,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testLineWidth120() {
 			var path = createBiomeConfig("biome/config/line-width-120.json");
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withConfigPath(path).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter120.js");
 		}
@@ -213,7 +239,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testLineWidth80() {
 			var path = createBiomeConfig("biome/config/line-width-80.json");
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withConfigPath(path).create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withConfigPath(path).create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/longLineBefore.js", "biome/js/longLineAfter80.js");
 		}
@@ -231,7 +257,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCjs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.cjs", "biome/js/fileAfter.cjs");
 		}
@@ -243,7 +269,7 @@ class BiomeStepTest extends ResourceHarness {
 		@Test
 		void testSetLanguageCssExperimental() {
 			var path = createBiomeConfig("biome/config/css-enabled.json");
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.8.3", downloadDir.toString()).withConfigPath(path).withLanguage("css").create();
+			var step = BiomeStep.withExeDownload("1.8.3", downloadDir.toString()).withConfigPath(path).withLanguage("css").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -254,7 +280,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCssStable() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.9.0", downloadDir.toString()).withLanguage("css").create();
+			var step = BiomeStep.withExeDownload("1.9.0", downloadDir.toString()).withLanguage("css").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/css/fileBefore.css", "biome/css/fileAfter.css");
 		}
@@ -265,7 +291,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageCts() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.cts", "biome/ts/fileAfter.cts");
 		}
@@ -276,7 +302,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.js", "biome/js/fileAfter.js");
 		}
@@ -287,7 +313,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJson() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("json").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("json").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/json/fileBefore.json", "biome/json/fileAfter.json");
 		}
@@ -298,7 +324,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJsonc() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("jsonc").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("jsonc").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/jsonc/fileBefore.jsonc", "biome/jsonc/fileAfter.jsonc");
 		}
@@ -309,7 +335,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageJsx() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("jsx").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("jsx").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.jsx", "biome/js/fileAfter.jsx");
 		}
@@ -320,7 +346,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageMjs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("js").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("js").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/js/fileBefore.mjs", "biome/js/fileAfter.mjs");
 		}
@@ -331,7 +357,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageMts() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.mts", "biome/ts/fileAfter.mts");
 		}
@@ -342,7 +368,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageTs() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("ts").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("ts").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.ts", "biome/ts/fileAfter.ts");
 		}
@@ -353,7 +379,7 @@ class BiomeStepTest extends ResourceHarness {
 		 */
 		@Test
 		void testSetLanguageTsx() {
-			var step = BiomeStep.withExeDownload(BiomeFlavor.BIOME, "1.2.0", downloadDir.toString()).withLanguage("tsx").create();
+			var step = BiomeStep.withExeDownload("1.2.0", downloadDir.toString()).withLanguage("tsx").create();
 			var stepHarness = StepHarnessWithFile.forStep(BiomeStepTest.this, step);
 			stepHarness.testResource("biome/ts/fileBefore.tsx", "biome/ts/fileAfter.tsx");
 		}